Refactored platform modulue. Platform subclasses reside in their own directory. Added universal selenium timeout constant
