+++ scripts/services/auto.py
@@ -31,7 +31,9 @@ class AutoService(object):
         if not self.gui_support:
             print("GUI not supported on this system, checking for additional dependencies")
             check_xvfb()
+            ####### DO NOT REMOVE CONDITIONAL IMPORT #######
             from pyvirtualdisplay import Display # Should be installed in check_xvfb()
+            ################################################
             self.display = Display(visible=0, size=(800, 600))
             self.display.start()
