+++ scripts/services/configuration.py
@@ -3,7 +3,8 @@ import sys
 import constants.constants as constants
 
 from selenium import webdriver
-from .platform import PlatformBuilder, Platform
+from ..models.platform import Platform
+from .platform import PlatformBuilder
 
 # Added to make the utils module available to the script
 sys.path.append(f"{pathlib.Path(__file__).parent.resolve()}/../..")
